+++ miden-crypto/src/ecdh/k256.rs
@@ -218,10 +218,11 @@ impl KeyAgreementScheme for K256 {
     fn extract_key_material(
         shared_secret: &Self::SharedSecret,
         length: usize,
+        info: &[u8],
     ) -> Result<Vec<u8>, super::KeyAgreementError> {
         let hkdf = shared_secret.extract(None);
         let mut buf = vec![0_u8; length];
-        hkdf.expand(&[], &mut buf)
+        hkdf.expand(info, &mut buf)
             .map_err(|_| super::KeyAgreementError::HkdfExpansionFailed)?;
         Ok(buf)
     }